miniscript: Inline finalizer module
The majority of the functions in the  `finalizer` module take a `Psbt`
as the first paramater, this means they can be implemented as methods on
`Psbt`. The remaining functions can be moved as well and made private to
the `miniscript` module.
